+++ src/scene/Camera.cpp
@@ -1,7 +1,7 @@
 #include "Camera.h"
 
 // Constructor with vectors
-Camera::Camera(glm::vec3 position, glm::vec3 up, float pitch, float yaw) : Front(glm::vec3(0.0f, 0.0f, -1.0f)), Pitch(pitch), Yaw(yaw)
+Camera::Camera(glm::dvec3 position, glm::dvec3 up, double pitch, double yaw) : Front(glm::dvec3(0.0f, 0.0f, -1.0f)), Pitch(pitch), Yaw(yaw)
 {
 	Position = position;
 	WorldUp = up;
@@ -9,9 +9,9 @@ Camera::Camera(glm::vec3 position, glm::vec3 up, float pitch, float yaw) : Front
 }
 
 // Processes input received from any keyboard-like input system. Accepts input parameter in the form of camera defined ENUM (to abstract it from windowing systems)
-void Camera::ProcessKeyboard(Camera_Movement direction, float deltaTime)
+void Camera::ProcessKeyboard(Camera_Movement direction, double deltaTime)
 {
-	float velocity = MovementSpeed * deltaTime;
+	double velocity = MovementSpeed * deltaTime;
 	if (direction == Camera_Movement::FORWARD)
 		Position += Front * velocity;
 	if (direction == Camera_Movement::BACKWARD)
@@ -27,7 +27,7 @@ void Camera::ProcessKeyboard(Camera_Movement direction, float deltaTime)
 }
 
 // Processes input received from a mouse input system. Expects the offset value in both the x and y direction.
-void Camera::ProcessMouseMovement(float xoffset, float yoffset, bool constrainPitch)
+void Camera::ProcessMouseMovement(double xoffset, double yoffset, bool constrainPitch)
 {
 	Yaw += xoffset * MouseSensitivity;
 	Pitch += yoffset * MouseSensitivity;
@@ -45,15 +45,15 @@ void Camera::ProcessMouseMovement(float xoffset, float yoffset, bool constrainPi
 	updateCameraVectors();
 }
 
-void Camera::ProcessJoystickMove(float x, float y, float zL, float zR, float deltaTime) {
-	float velocity = MovementSpeed * deltaTime * joystickMoveAccel;
+void Camera::ProcessJoystickMove(double x, double y, double zL, double zR, double deltaTime) {
+	double velocity = MovementSpeed * deltaTime * joystickMoveAccel;
 	Position += Front * x * velocity;
 	Position += Right * y * velocity;
 	Position -= Up * zL * velocity;
 	Position += Up * zR * velocity;
 }
 
-void Camera::ProcessJoystickLook(float x, float y, float deltaTime) {
+void Camera::ProcessJoystickLook(double x, double y, double deltaTime) {
 	Yaw += x * joystickLookAccel*deltaTime;
 	Pitch += y * joystickLookAccel*deltaTime;
 
@@ -70,7 +70,7 @@ void Camera::ProcessJoystickLook(float x, float y, float deltaTime) {
 }
 
 // Processes input received from a mouse scroll-wheel event. Only requires input on the vertical wheel-axis
-void Camera::ProcessMouseScroll(float yoffset, float deltaTime)
+void Camera::ProcessMouseScroll(double yoffset, double deltaTime)
 {
 	//if (Zoom >= 1.0f && Zoom <= 45.0f)
 	//	Zoom -= yoffset;
@@ -86,7 +86,7 @@ void Camera::ProcessMouseScroll(float yoffset, float deltaTime)
 	
 }
 
-void Camera::ChangeCameraSpeed(Camera_Movement direction, float deltaTime) {
+void Camera::ChangeCameraSpeed(Camera_Movement direction, double deltaTime) {
 	if (direction == Camera_Movement::UP)
 		MovementSpeed += 3 * MovementSpeed * deltaTime;
 	if (direction == Camera_Movement::DOWN)
